More API work on the list component
