+++ Examples/WildeGuess/WildeGuess/Components/LoadingIndicatorComponent.h
@@ -1,0 +1,15 @@
+//
+//  LoadingIndicatorComponent.h
+//  WildeGuess
+//
+//  Created by Oliver Rickard on 2/4/17.
+//
+//
+
+#import <ComponentKit/CKCompositeComponent.h>
+
+@interface LoadingIndicatorComponent : CKCompositeComponent
+
++ (instancetype)new;
+
+@end
